Fix several codeql warns

+++ src/chipset/scamp.c
@@ -840,7 +840,7 @@ recalc_sltptr(scamp_t *dev)
             recalc_ems(dev);
         }
     } else {
-        for (uint8_t i = 0; i < (sltptr / EMS_PGSIZE); i++)
+        for (uint32_t i = 0; i < (sltptr / EMS_PGSIZE); i++)
             scamp_mem_update_state(dev, i * EMS_PGSIZE, EMS_PGSIZE, 0x00, MEM_FMASK_SLOTBUS);
 
         for (uint8_t i = (sltptr / EMS_PGSIZE); i < 40; i++)
